Split arena module into 2 submodules Fix imports
