+++ _mongo/user.js
@@ -1,0 +1,8 @@
+db = db.getSiblingDB('arena')
+db.createUser(
+    {
+        user: _getEnv('MONGODB_USER'),
+        pwd: _getEnv('MONGODB_PASSWORD'),
+        roles: [ { role: "readWrite", db: "arena" } ]
+    }
+)
