+++ bot/commands/gladiator/show.go
@@ -7,7 +7,7 @@ import (
 	"github.com/diamondburned/arikawa/v3/discord"
 	"github.com/diamondburned/arikawa/v3/gateway"
 	"github.com/diamondburned/arikawa/v3/utils/json/option"
-	"github.com/wihrt/idle_arena/arena"
+	"github.com/wihrt/idle_arena/arena/client"
 	"github.com/wihrt/idle_arena/bot/utils"
 	"go.uber.org/zap"
 )
@@ -16,11 +16,11 @@ func ShowGladiatorsMenu(e *gateway.InteractionCreateEvent) (api.InteractionRespo
 	var (
 		mID  = utils.GenerateManagerID(e)
 		url  = os.Getenv("ARENA_URL")
-		a    = arena.NewClient(url)
+		c    = client.NewClient(url)
 		data api.InteractionResponse
 	)
 
-	g, err := a.GetGladiators(mID)
+	g, err := c.GetGladiators(mID)
 	if err != nil {
 		zap.L().Error("Cannot get gladiators",
 			zap.String("managerID", mID),
@@ -47,7 +47,7 @@ func ShowGladiators(e *gateway.InteractionCreateEvent) (api.InteractionResponse,
 	var (
 		mID    = utils.GenerateManagerID(e)
 		url    = os.Getenv("ARENA_URL")
-		a      = arena.NewClient(url)
+		a      = client.NewClient(url)
 		data   api.InteractionResponse
 		eArray []discord.Embed
 	)
