+++ bot/commands/manager/retire.go
@@ -6,7 +6,7 @@ import (
 	"github.com/diamondburned/arikawa/v3/api"
 	"github.com/diamondburned/arikawa/v3/gateway"
 	"github.com/diamondburned/arikawa/v3/utils/json/option"
-	"github.com/wihrt/idle_arena/arena"
+	"github.com/wihrt/idle_arena/arena/client"
 	"github.com/wihrt/idle_arena/bot/utils"
 	"go.uber.org/zap"
 )
@@ -15,11 +15,11 @@ func RetireManager(e *gateway.InteractionCreateEvent) (api.InteractionResponse, 
 	var (
 		mID  = utils.GenerateManagerID(e)
 		url  = os.Getenv("ARENA_URL")
-		a    = arena.NewClient(url)
+		c    = client.NewClient(url)
 		data api.InteractionResponse
 	)
 
-	err := a.RetireManager(mID)
+	err := c.RetireManager(mID)
 	if err != nil {
 		zap.L().Error("Cannot retire manager",
 			zap.String("UserID", e.Member.User.ID.String()),
