+++ docker-compose.yml
@@ -6,7 +6,7 @@ services:
             context: .
             args:
                 GOBINARY: arenaBot
-                GOFILE: ./bot/bot.go
+                GOFILE: ./bin/bot/bot.go
         container_name: bot
         environment:
             APP_ID: $APP_ID
@@ -22,10 +22,10 @@ services:
             context: .
             args:
                 GOBINARY: arenaBackend
-                GOFILE: ./backend/backend.go
+                GOFILE: ./bin/backend/backend.go
         container_name: backend
         environment:
-            MONGO_URL: mongodb://$MONGODB_USER:$MONGODB_PASSWORD@mongo:27017/$MONGODB_DATABASE?authSource=admin
+            MONGO_URL: mongodb://$MONGODB_USER:$MONGODB_PASSWORD@mongo:27017/$MONGODB_DATABASE
             HTTP_PORT: $BACKEND_PORT
         entrypoint: ["/usr/bin/arenaBackend"]
         ports:
@@ -39,10 +39,10 @@ services:
             context: .
             args:
                 GOBINARY: arenaLoader
-                GOFILE: ./loader/loader.go
+                GOFILE: ./bin/loader/loader.go
         container_name: loader
         environment:
-            MONGO_URL: mongodb://$MONGODB_USER:$MONGODB_PASSWORD@mongo:27017/$MONGODB_DATABASE?authSource=admin
+            MONGO_URL: mongodb://$MONGODB_USER:$MONGODB_PASSWORD@mongo:27017/$MONGODB_DATABASE
             DND_API: https://www.dnd5eapi.co
             TIMEOUT: 30s
             INDEXES: "armor,weapon"
@@ -50,23 +50,24 @@ services:
         networks:
             - arenaBack
     mongo:
-        image: mongo:latest
+        image: bitnami/mongodb:latest
         container_name: mongo
         environment:
-            MONGO_INITDB_ROOT_USERNAME: $MONGODB_ROOT_USER
-            MONGO_INITDB_ROOT_PASSWORD: $MONGODB_ROOT_PASSWORD
-            MONGODB_INIT_DATABASE: $MONGODB_DATABASE
+            MONGODB_ROOT_PASSWORD: $MONGODB_ROOT_PASSWORD
+            MONGODB_DATABASE: $MONGODB_DATABASE
             MONGODB_USER: $MONGODB_USER
             MONGODB_PASSWORD: $MONGODB_PASSWORD
         networks:
             - arenaBack
             - arenaAdmin
+        volumes:
+            - ./_mongo:/docker-entrypoint-initdb.d/
         restart: always
     mongo-express:
         image: mongo-express:latest
         container_name: mongo-express
         environment: 
-            ME_CONFIG_MONGODB_URL: mongodb://$MONGODB_USER:$MONGODB_PASSWORD@mongo:27017/$MONGODB_DATABASE?authSource=admin
+            ME_CONFIG_MONGODB_URL: mongodb://$MONGODB_USER:$MONGODB_PASSWORD@mongo:27017/$MONGODB_DATABASE
         ports:
             - 8081:8081
         networks:
